+++ src/functions/municipality.ts
@@ -1,6 +1,9 @@
-import { getIndexedMunicipalitiesByProvince } from '../utils/data-loader';
+import {
+  getIndexedMunicipalitiesByCode,
+  getIndexedMunicipalitiesByProvince,
+  getMunicipalities,
+} from '../utils/data-loader';
 import { PHMunicipality } from '../types/municipality';
-import { sortByName } from '../utils/sort';
 
 /**
  * Retrieves municipalities filtered by the specified province name and sorts them alphabetically.
@@ -9,8 +12,19 @@ import { sortByName } from '../utils/sort';
  * @returns {ReadonlyArray<PHMunicipality>} An array of municipalities belonging to the specified province, sorted alphabetically.
  */
 function getMunicipalitiesByProvince(code: string): readonly PHMunicipality[] {
-  const municipalities = getIndexedMunicipalitiesByProvince().get(code) || [];
-  return sortByName(municipalities);
+  return getIndexedMunicipalitiesByProvince().get(code) || [];
 }
 
-export { getMunicipalitiesByProvince };
+function getAllMunicipalities(): readonly PHMunicipality[] {
+  return getMunicipalities();
+}
+
+function getMunicipalityByCode(code: string): PHMunicipality | undefined {
+  return getIndexedMunicipalitiesByCode().get(code);
+}
+
+export {
+  getAllMunicipalities,
+  getMunicipalityByCode,
+  getMunicipalitiesByProvince,
+};