+++ src/functions/region.ts
@@ -1,1 +1,10 @@
-export { getRegions as getAllRegions } from '../utils/data-loader';
+import { getIndexedRegionsByCode, getRegions } from '../utils/data-loader';
+import { PHRegion } from '../types/region';
+
+const getAllRegions = getRegions;
+
+function getRegionByCode(code: string): PHRegion | undefined {
+  return getIndexedRegionsByCode().get(code);
+}
+
+export { getAllRegions, getRegionByCode };