+++ src/index.ts
@@ -1,12 +1,38 @@
-import { getBarangaysByMunicipality } from "./functions/barangay";
-import { getMunicipalitiesByProvince } from "./functions/municipality";
-import { getAllProvinces, getProvincesByRegion } from "./functions/province";
-import { getAllRegions } from "./functions/region";
+import {
+    getAddressByBarangayCode,
+    getAllBarangays,
+    getBarangayByCode,
+    getBarangaysByMunicipality,
+} from "./functions/barangay";
+import {
+    getAllMunicipalities,
+    getMunicipalityByCode,
+    getMunicipalitiesByProvince,
+} from "./functions/municipality";
+import {
+    getAllProvinces,
+    getProvinceByCode,
+    getProvincesByRegion,
+} from "./functions/province";
+import { getAllRegions, getRegionByCode } from "./functions/region";
+
+export type { PHBarangay } from "./types/barangay";
+export type { PHAddress } from "./types/address";
+export type { PHMunicipality } from "./types/municipality";
+export type { PHProvince } from "./types/province";
+export type { PHRegion } from "./types/region";
 
 export {
+    getAddressByBarangayCode,
+    getAllBarangays,
+    getAllMunicipalities,
     getAllProvinces,
     getAllRegions,
+    getBarangayByCode,
     getBarangaysByMunicipality,
+    getMunicipalityByCode,
     getMunicipalitiesByProvince,
-    getProvincesByRegion
+    getProvinceByCode,
+    getProvincesByRegion,
+    getRegionByCode,
 };